ibmcloud/requirements.txt
slackclient
softlayer
requests
pyyaml
